+++ packages/backend/jest.config.js
@@ -1,4 +1,8 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  automock: false,
+  setupFiles: [
+    "./setupJest.ts"
+  ]
 };