+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Recent and upcoming changes to lightdash
 ### Added
  - Add options menu for each table/dimension/metric
  - Add dialog with the relative file path and code relevant for each table/dimension/metric
+ - Add footer with Lightdash version
+ - Add Lightdash about dialog
+ - Warn user when there is a new version available
 
 ### Fixed
  - UI bug where the search box in the sidebar was cutoff